Remove the need for `exclude_end` in TextViewWidget/Caret!

+++ gui/renderer/text_renderer.cc
@@ -130,7 +130,7 @@ void TextRenderer::renderLineLayout(const font::LineLayout& line_layout,
 
             // Invert glyph y-offset since we're flipping across the y-axis in OpenGL.
             Vec4 glyph_copy = rglyph.glyph;
-            glyph_copy.y = static_cast<float>(metrics.line_height) - glyph_copy.y;
+            glyph_copy.y = metrics.line_height - glyph_copy.y;
 
             const Vec2 coords_vec = {
                 .x = static_cast<float>(glyph_coords.x),
